Revert depth d again; discrepancy happens randomly for unknown reasons
A shortcut to test this out is to repeatedly run the same code;
there is no discrepancy if and only if the contig values, e.g. heuristic
values are the same.

In this case, running the same code gave different values for the
contigs, which means that the results are going to be different.

As the algorithm is supposed to be deterministic, this is an anomaly.
It could be other assembly files, i.e. assembly, checkpoint and decode
folders, interfering with the decoding process. However, more
investigation is needed to confirm the cause of this discrepancy, and
why this may cause the issue.