Generate coeffs for random search in parser, standardize seed in params
Random search used to generate coeffs each time the algorithm is run;
this is redundant, and coeffs are now generated in
parse_args_based_on_strategy, depending on what is typed in by the user.
With the polynomial coefficients, a polynomial lambda function is
created, so algorithm now uses it as a parameter; because of this, it is
now identical to weighted random search.

parse_seed now always put seed into parameters, so that seed used is
always from parameters (rather than checking if it exists in
parameters every time it is needed). Default seed (if not specified by
the user) now is the seed specified in hyperparameters.