Parse seed in deterministic strategies too, optimize weighted_random
Seed is also used in the selection of the starting positions, including
deterministic strategies. This also simplifies the code as the seed is
now always in parameters.

The function f (in weighted_random) is now applied to edge probabilities
in get_contigs_greedy; this eliminates repeated computations of f(p) in
weighted_random_search, thus reducing runtime.