Add flag that uses function in hyperparams for weighted random
Default value of coeffs is seemingly changed to None; it is a sentinel
value to indicate if its has been specified (if not specified, it will
be None).

This new flag cannot be used with --coeffs as they are
contradictory - if it were to be allowed, it may be confusing to users
as to how the behaviour should be, so a way to avoid this ambiguity is
to not allow it (by raising an Exception).

The function in hyperparameters.py has been added accordingly.